+++ Source/Dictionary/Values/DictionaryUIntValue.swift
@@ -8,11 +8,11 @@ public struct DictionaryUIntValue: DictionaryValue {
     }
     
     public func serialize(src: any DictionaryKeyTypes, builder: Builder) throws {
-        guard let src = src as? UInt32 else {
+        guard let src = src as? UInt64 else {
             throw TonError.custom("Wrong src type. Expected uint32")
         }
         
-        try builder.storeUint(UInt64(src), bits: bits)
+        try builder.storeUint(src, bits: bits)
     }
     
     public func parse(src: Slice) throws -> any DictionaryKeyTypes {
